+++ src/test/java/org/gridsuite/modification/server/modifications/VscModificationTest.java
@@ -30,7 +30,7 @@ import java.util.stream.IntStream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.gridsuite.modification.server.NetworkModificationException.Type.WRONG_HVDC_ANGLE_DROOP_ACTIVE_POWER_CONTROL;
-import static org.gridsuite.modification.server.modifications.ModificationUtils.ACTIVE_POWER_CONTROL_DROOP_P0_REQUIRED_ERROR_MSG;
+import static org.gridsuite.modification.server.modifications.VscModification.ACTIVE_POWER_CONTROL_DROOP_P0_REQUIRED_ERROR_MSG;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
@@ -243,39 +243,6 @@ public class VscModificationTest extends AbstractNetworkModificationTest {
         Assert.assertEquals(1, activePowerControl.getDroop(), 0);
         Assert.assertTrue(activePowerControl.isEnabled());
     }
-
-//    @ParameterizedTest(name = "Test modification with absent droop infos when extension does not yet exist [{index}] {0}, {1}, {2}")
-//    @CsvSource({
-//        "true,false,false",
-//        "true,true,false",
-//        "true,false,true",
-//        "false,true,false",
-//        "false,true,true",
-//        "false,false,true",
-//    })
-//    void testHvdcAngleDroopActivePowerControlWithAbsentInfos(boolean isPresentAngleDroopActivePowerControl, boolean isPresentDroop, boolean isPresentP0) {
-//        var networkuuid = UUID.randomUUID();
-//        Network networkWithoutExt = NetworkCreation.createWithVSC(networkuuid, false);
-//        VscModificationInfos modificationInfos = (VscModificationInfos) buildModification();
-//
-//        // reset null depending to test arguments
-//        if (!isPresentAngleDroopActivePowerControl) {
-//            modificationInfos.setAngleDroopActivePowerControl(null);
-//        }
-//        if (!isPresentDroop) {
-//            modificationInfos.setDroop(null);
-//        }
-//        if (!isPresentP0) {
-//            modificationInfos.setP0(null);
-//        }
-//
-//        VscModification wrongVscModification = new VscModification(modificationInfos);
-//        String message = Assert.assertThrows(NetworkModificationException.class,
-//                () -> wrongVscModification.check(networkWithoutExt))
-//            .getMessage();
-//        assertThat(message).isEqualTo(WRONG_HVDC_ANGLE_DROOP_ACTIVE_POWER_CONTROL.name() + " : "
-//              + ACTIVE_POWER_CONTROL_DROOP_P0_REQUIRED_ERROR_MSG);
-//    }
 
     @Test
     public void testAngleDroopActivePowerControlWithAbsentInfos() {
